+++ src/lib/components/sailor/dialogs/SelectDialog.svelte
@@ -37,7 +37,7 @@
 </script>
 
 <Dialog.Root bind:open>
-  <Dialog.Content class="sm:max-w-[425px]">
+  <Dialog.Content>
     <Dialog.Header>
       <Dialog.Title>{title}</Dialog.Title>
       {#if description}